feat[CalcModel]: new factory class calculator to hide concrete classes
The CalcModel class implements a Factory method which the client calls
to calculate an expression. The client sends a String expression to be
calculated and a boolean stating if the client wants a infix or postfix
calculation and the Factory class calls the concrete calculator
implementations evaluate() method to return a answer float. The answer
is then stored in a memory Stack and returned to the client.